Reorganize TestBuilder functions
